data: remove un-needed check with datavar statistic
The Chi2DataVar statistic used to error out if given a counts
value of 0, so the get_staterror() method added a workaround for
the background case: if any background group had 0 counts when
the datavar statistic was in use it would ignore the contribution
to the error from this background group. However, in #153 the
datavar behavior was changed to only error out if negative values
were found, and a value of 0 would get a 0 error in it. This
means, now that we have added a check for this exact situation,
that we can remove the work-around as it isn't needed.