Tests: fix the random seed for a fake_pha test
This is the last test in test_fake_pha.py that had not fixed the
random seed. After it failing once-too-many times in a CI run I
decided to fix the seed to make this a repeatable test. The
numbers used are a regression test - that is, the counts that
get created - but it's hard to come up with a reliable test while
allowing the values to vary.

I did add an explicit check that, by adding the background model,
we get a vastly-different result (since without it the source model
evaluates to 0 for all bins).