stats: one way to fix #356
This is the simplest way I can see to address the XSPEC variance
issue, but it is not an ideal fix. Unfortunately, it is
likely that an ideal fix requires re-working a number of
assumptions made in the data and stats objects, and at this
time it is likely that we need to look at this to support
polarimetry X-ray data, which has different needs again.

The problem is that when Chi2XspecVar is used to calculate
the error for PHA data, the background is subtracted, and
one or both of the source or background has 0 counts in a
group, then the calculated error is larger than the XSPEC
version. This is because the error calculation is done by
sherpa.stats._statfcts.calc_chi2xspecvar_errors and at
this point the source and background arrays are being handled
individually so each count=0 value ends up with an error=1
value, and the source and background elements are then
combined. Unfortunately, XSPEC does this in a different order
and calculates the variance knowning that it has both source
and background components, and so handles things differently,

A)

The change is only made to the get_staterror method of the
sherpa.astro.data.DataPHA class, and not to the error routines
in sherpa.stats. Given the current design, this is actually
really the only place to do this as we do not just have to
worry about the counts in the source and background bins, but
we also have to care about the scaling values (exposure time,
backscal, and areascal) when calculating the replacement error.

B)

It is *not* a complete fix because there are ways to call the
XSPEC variance routine which will not trigger this fix, but
that is only if the user is doing something rather "clever" and
will not be a problem for the people we want to get this fix. The
issue here is the check in DataPHA.get_staterror for

    if staterrfunc == Chi2XspecVar.calc_staterror ...

Technically a user could end up calling this with staterrfunc
set to sherpa.stats._statdcts.calc_chi2xspecvar_errors which
should trigger the same behavior (i.e. the same fix) but I have
not included this. This is something that could be changed but I
did not feel like it was worth it.