Do not use options.data_files to refer to the *rc files
The setuptools document at
https://setuptools.pypa.io/en/latest/userguide/declarative_config.html#options
points out that options.data_files is deprecated, and we should
do something else, but - as with everything python-packaging
related - when I go to the "read here how to do it" page I do not
understand what they are saying. It looks like I can just use
options.package_data to refer to the sherpa[-standalone].rc files,
but it triggers build issues. This appears to be because

a) helpers/sherpa_config.py implicitly requires on the data_files
   option (one issue in doing this is understanding how a
   configuration option maps to various fields in the setuptools
   Command object, how thy are meant to be used, and how we are
   using them)

b) the reason we use the data_files field is to include the
   stack and group "modules" if needed (primarily if in a non-CIAO
   build), and this is something we've wanted to address for a
   long time:
   https://github.com/sherpa/sherpa/issues/50