helpers: use super to reference parent class
Instead of explicitly mentioning the parent class, use super().

There is also a small change in the develop routine so that

a) we announce the stk and group extensions separately
b) we only announce them when they are actually built

However it looks like we don't see these messages by default,
so most users will not see any difference.