helpers: remove unused code in sdist
I had noted that sdist used numpy.distutils.misc_util.get_data_files
and what should be the replacement for this. In testing, I've found
that I can't trigger the add_defaults method which calls this
routine, so I have replaced it with a method that errors out if
called (in case my testing, which has been limited, was not
sufficient).