helpers: always include the XSPEC package
Prior to this commit the XSPEC relevant code and tests were
dropped when XSPEC support was disabled. This made sense, but
there have been times when the installation could still install
the XSPEC tests, with the result that 'import sherpa.astro.xspec'
would not fail even though XSPEC support was not enabled. This
behavior seems to have been fixed with the PEP 517 changes, but it
is not-at-all-clear why it happened, so it seems better to guard
against this feature of the build infrastructure. See
https://github.com/sherpa/sherpa/issues/1545
for more information.

Instead, we now ensure the XSPEC code is installed, even when XSPEC
support is disabled. This can lead to an unsettling error message
from 'import sherpa.astro.xspec', so the XSPEC initialization code
has been tweaked to provide a nicer message in this case:

% python -c 'import sherpa.astro.xspec'
Traceback (most recent call last):
  File "/home/dburke/miniconda3/envs/sherpa-pr/lib/python3.9/site-packages/sherpa/astro/xspec/utils.py", line 24, in <module>
    from . import _xspec
ImportError: cannot import name '_xspec' from partially initialized module 'sherpa.astro.xspec' (most likely due to a circular import) (/home/dburke/miniconda3/envs/sherpa-pr/lib/python3.9/site-packages/sherpa/astro/xspec/__init__.py)

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/dburke/miniconda3/envs/sherpa-pr/lib/python3.9/site-packages/sherpa/astro/xspec/__init__.py", line 113, in <module>
    from .utils import ModelMeta, version_at_least, equal_or_greater_than
  File "/home/dburke/miniconda3/envs/sherpa-pr/lib/python3.9/site-packages/sherpa/astro/xspec/utils.py", line 31, in <module>
    raise ImportError("XSPEC support is not enabled") from ie
ImportError: XSPEC support is not enabled

Note that we retain the original "hard to understand for most users"
error, but only as a "secondary" one - the main error tells the
user the problem. We keep the original error in case there is a
problem in XSPEC initialization, which we'd like to know about.