helpers: move back to distutils for sdist
Looking cosely at the message from NumPy, it looks like we have
to use distutils after all. In doing so, we trigger the add_defaults
behavior. In checking out the old method, it was only being used
to add in the stack and group dynamic libraries, and this doesn't
actually work (they get rejected). So I removed the add_defaults
method.

If #50 gets fixed up we may have to do something, but I don't
think so because they are both compiled code and the sdist is not
for those files (at least as far as I know).