Allow group/stk to be accessed from the sherpa namespace
Issue #50 notes that we install two files outside the Sherpa
namespace - namely stk.so and group.so. This is because in the CIAO
installation we take these from the CIAO environment (where they
are not in a namespace). One way to address #50 is to allow these
two files to be imported from within Sherpa (a standalone
installation) or from CIAO. This commit supports this, but does
not make any change in how the code is built (so it doesn't address
the underlying issue).

The code is basically

    try:
        from sherpa import <name>
    except ImportError:
        import <name>

and we could swap the order (i.e. try the non-namespace version
first), which would be *marginally* faster for CIAO users, but
I think it's better to try the namespaced version first, and
the runtime difference is not going to be an issue given the large
number of modules that are being loaded.