Tests: switch to pytest.approx for an XSPEC test
The test_evaluate_xspec_model test already used pytest.approx but
for some reason test_evaluate_xspec_model_noncontiguous2 was still
using assert_allclose. We now use pytest.approx.

This was done in part because testing with XSPEC 12.12.1 failed
for the ismabs and olivineabs models for the non-contiguous case,
which was down to default tolerances: assert_allclose was using
a relative tolerance of 1e-7 which failed but if we bumped it to
1e-6 the test would pass (the pytest.approx defaults do not need
changing).

I will note that for these two models a combination of the
energy/wavelength range and the default XSPEC abundances/whatever
mean that this test is not particularly useful (the calculated
values are not really realistic), but at least the results are
repeateable.