Update the RTD build for recent changes
RTD is using Sphinx 5 which complains if language is set to None.

The python version has been bumped. Note that we need to update
the RTD configuration to use the latest configuration system,
but that can wait.

The main change is to ensure that matplotlib is mocked, as that is
what we need to get the docs to build. I have dropped the
errorbar over-ride, which should not be needed anymore due to
the decoupling of options into backend-specific and generic in
this PR series.