ui: report filter after notice/ignore
Report the filter expression after the user has noticed
or ignored something. This is done at the ui level rather
than when calling a method on the Data object, as it is
felt to be more-useful there. It could be done on the data
object, but that is a bit-more "magic" (i.e. the method
has side effects) than I think is warranted.

There is some annoying repetition of code in the astro/2D
case because we do not allow arbitrary kwargs, unlike the
1D case. This could be addressed here but I think it best
left for another PR.

The difference between 1D and 2D filtering (e.g. what does
"" mean, is it 'no data' or 'all data', as understood by
1D and 2D data respectively at the moment) complicates this
code but sorting this out is for a different PR. See #1430
for more information.

Handling of the DataPHA case, in particular the background
files, also complicates things.

There is some ambiguity in how we express filters - e.g.
do we use ':' or '-' to separate ranges - which is not
supported when we parse filters (e.g. parse_expr requires
the range to be separated by ':'). We should look at fixing
this (probably by only allowing ':' to separate expressions),
but that is also for another PR.

Some tests are broken by this - these are tests for which
ignore_bad (which we nkow to be problematic) has issues,
so we treat them as regression tests, and just allow them
to behave as they do. Once we fix the "bad" issues we should
then notice the tests failing and can address them then.

This addresses #1558.