Tests: fix a logical test error
The test should have used "m1" to reference the model component
to plot, but I'd accidentally used "mdl" (as this is the local
name used in the code for some reason). The problem is that we
can persist model names over different test set-ups, so this was
never noticed (as we don't really check the plot results in depth
here).

In PR #1537 I added code to try and avoid the underlying problem -
that is, model names can remain even is we have throw-away
sessions - but it appears to not work here. I noticed this test
issue when I was running the tests in parallel, and that obviously
"hit the right spot" for running the plot tests without magically
creating a component called "mdl" (a name we use a lot in the
tests).

The fix is easy, I just wanted to provide some context, as it
appears to be related to issue #1527.