conftest: try to avoid issue #1590
From a numbre of searches, it seems like the

    Tcl_AsyncDelete: async handler deleted by the wrong thread

error "caused" by matplotlib, as it defaults to the TkAgg backend.
If we switch to the Agg backend does this fix the problem?

Note that we currently don't actually test the matplotlib output so
there should be no prioblem in the tests.