data: remove numpy array conversion warnings
Remove the user-warning messages you would get from sending in a
non-ndarray to the independent and dependent axes. So, instead of

  >>> d = Data1D('x', [1, 2, 3], [4, 5, 6])
/home/dburke/miniconda3/envs/ciao414/lib/python3.9/site-packages/sherpa/data.py:88: UserWarning: Converting array [4, 5, 6] to numpy array.
  warnings.warn("Converting array {} to numpy array.".format(array))

  >>> d = DataPHA('x', [1, 2, 3], [4, 5, 6])
/home/dburke/miniconda3/envs/ciao414/lib/python3.9/site-packages/sherpa/data.py:88: UserWarning: Converting array [1, 2, 3] to numpy array.
  warnings.warn("Converting array {} to numpy array.".format(array))
/home/dburke/miniconda3/envs/ciao414/lib/python3.9/site-packages/sherpa/data.py:88: UserWarning: Converting array [4, 5, 6] to numpy array.
  warnings.warn("Converting array {} to numpy array.".format(array))

you now get no warning.

The reason is that we now make it more obvious that the argument
has been converted (e.g. we try to always create a ndarray-compatible
value for the fields).

We do still warn for certain mask cases, but this is a much-rarer
condition than just sending in a list to Data1D.