ui: impove notice/ignore behavior in case of errors
For various reasons we may get an error when restring the filter
expression. As we don't want this to derail the notice or ignore
command (even if it indicates that something is wrong, it may not
be anything that affects the user), then try to catch all the
known issues. We could just catch all errors, but let's try to
be surgical.

The handling of ignore_bad is interesting, as it causes the filter
to be displayed, but it's not obvious whether it ever will cause
an expression to change.

The Data2D case is annoying, as the ui ignore/notice layer is not
designed to handle this type of data (see #1587). We notice it
here in two places

a) no get_xlabel method, so we have to catch it and call
   get_x0label / get_x1label instead

b) even with that, no filter is ever displayed because

   1. the get_filter method always returns ''
   2. but we don't get this far because the method does not have
      any kwargs, so when we call it with delim=':' then the call
      fails, which we trap and so do not generate any output

There's also some changes made to try and remove the knowledge
of the bkg_id argument from the UI code. At the moment this has
just moved the logic around (and made it repeated in a number of
cases), but at least it is more obvious now and may be simplified
in later work.