ui: improve notice2d/ignore2d output
This is related to issue #1430 and is related to the fact that
for DataIMG the get_filter output is the empty string for both
"all data is included" and "all data is excluded". We now use the
mask field to identify these two conditions, and only call
get_filter when data is partially included.

A number of tests have been tweaked:

- simplify the testing of the caplog output, to avoid quite so
  much boilerplate text
- explicitly avoid the caplog messages we don't care about
- add some explicit tests of the ignore caplog output (related to
  ignore_bad handling)