ui: when the mask is False we know all data has been filtered
This follows the DataIMG case, where we can use the mask attribute
to find out if all the data has been ignored. Doing so turns out
to remove a special-case for the DataPHA code, where with mask
set to False we still got a string message. Ideally this would be
addressed in the Data code, but this is a simple fix.

I had hoped that we could avoid some of the exception checks, but
we can not be sure we won't hit one of these problem cases (at
least without doing a proper analysis of the ignore/notice code).

I had also hoped that we could handle the mask is True case, but
we don't have a 1D equivalent to the "Field()" filter.