Address review comments
An explicit test of the notice/ignore output from a PHA file,
and with wavelength analysis, has been added.

The notice and notice_id code has been simplified, with the
logic now being in notice_data_range. If this were a method
of the Session subclass we could probably move the PHA
background-handlnig code there (but would then still need to
work out what to do with the documentation as it's currently
defined in the general class, not the astro one).

Co-authored-by: Hans Moritz Günther <moritz.guenther@gmx.de>