Improvements for doctests, skip if datadir not found
This bundles a number of separate improvements.
We now use sherpa functionality to find the directory where data files
that may be needed by doctests are do be found; if those are not found, we skip
all doctests with an obvious message. While in principle, it woudl be possible
to run the remaining doctests, I felt this was a cleaner appraoch.
Since doctest will essentially only be run by developers, we cna expect that
the test data will be present in most cases.

I also found that the default for floating point comparison was too strict
for tests that run fits. For simplicity I've loosened that requirement
globally. While many tests can be done to higher accuracy, the point of
doctests is to not clutter the docstrings or rst files; more detailed tests
are for real unit test written as code.