Tests: add a test for histogram1/2d read-only axes
With #1477 the output of get_dep is marked read-only, which means
that the second example for histogram1d now fails. Add tests for
both 1d and 2d histogram variants that check this condition, and
tweak the histogram1d example to be a bit easier to follow (and,
in future, test).