utils: ensure histogram1d/2d copy the input arguments before sorting
Ensure we copy the arguments before sorting them. It would
probably be better to just error out if the limits aren't
sorted for the 1D case, but that's for future discussion.

I note that get_error_estimates can also sort an input argument,
but there it is possibly okay (or implied) by the documentation,
so no change is made.