Explicitly include setup.py in MANIFEST.in
This was needed when testing out tox, so pull it out into a separate
commit. I do not understand the packaging ecosystem enough to know if
this was a temporary issue (e.g. one particular setuptools release) or
an actual problem. It seems it should not be a problem to explicitly
include setup.py, at least as long as we still require it.