Rework XSPEC state settings in set_xsstate
We have to decide how to handle "missing" settings from the
input sent to set_xsstate. Previously we had an interesting check,
that basically checked we had all settings, and then if any were
missing we did nothing. This makes the code a bit awkward to
handle potential updates to the state handling (e.g. if we want
to also track XFLT keywords), and to handle state files stored
with different Sherpa versions.

The code has now been changed to treat each setting independently,
rather than an all-or-nothing change.