xspec: rework XSPEC state settings in set_xsstate
We have to decide how to handle "missing" settings from the
input sent to set_xsstate. Previously we had an interesting check,
that basically checked we had all settings, and then if any were
missing we did nothing. This makes the code a bit awkward to
handle potential updates to the state handling (e.g. if we want
to also track XFLT keywords), and to handle state files stored
with different Sherpa versions.

The code has now been changed to treat each setting independently,
rather than an all-or-nothing change.

As part of this, and to support some planned additional features,
the state tracking that is done by XSPEC - which is necessary
because the XSPEC model API does not always provide a way to find
out "what the user has changed" - has been changed to use a single
top-level dictionary, rather than a collection of top-level
variables. As it is not entirely clear what the best approach is,
and it is likely to change as XSPEC changes, using an untyped
dictionary, rather than something like a dataclass or other way
to provide structure, seems appropriate.