xspec: get_xsxset return dict, clear_xsxset, error handling
Add a way to clear the XSET database - we already have one,
matching XSPEC, but it's not been advertised to the user, so
it makes sense to add a simple way to do it.

Allow the full database to be returned, by calling get_xsxset
with no argument. This requires C++11 support.

We now raise an error if get_xsxset is called with an invalid
keyword, rather than returning an empty string, to better
match Python dictionaries.