xspec: add support for set_xspath_model
We can set the model path as well as the manager path, so do so.

This introduces the ability to send in a path as an actual Path
object, rather than just a string. There's a lot of discussion
about "how much Path is too much", but here

- these are not often used routines
- the overhead of supporting Path as well as str is low

There is the thought that we could return a Path rather than a
str, but that would be a larger, breaking change.

For get/set_xsstate, we could avoid tracking the paths settings,
but then we can not identify the difference between the user
explicitly setting the value to this is the default value. So we
track the values in xsstate["paths"].