xspec: initial support for read/write XFLT keywords.
There is no way to make use of these keywords at the moment
as we can not send in the spectrumNumber argument to XSPEC models
(other than using dataset 1, which we default to using for all
model calls). Note that this code requires C++11.

Since information is normally provided by the FITS header,
make `load_xsxflt` be the primary means of setting the values,
although users who understand the models can use the lower-level
routines like `set_xsxflt` themselves. However, the idea is that
users have all the settings that are needed rather than providing
an interface that can be added to one at a time (at least, for
an individual spectrumNumber). This just avoids having to deal
with "is this dict empty because we are waiting for more data or
because it's empty" type of questions.