xspec: return the element table used by XSPEC
Rather than provide a per-element routine, this just returns a dict
with all elements (indexed by name). We also add a routine to return
the number of elements.

This could have been implemented as just creating a Python dict
with the value {'H': 1, 'He': 2, ..., 'Zn': 30}, but we use the
XSPEC interface so that we will handle the rather-unlikely case
that XSPEC adds more elements.

Internally this adds the irange class to try and use C++ iterators
to replace a simple for(i=0; i<N; ++i) statement. This could be
extended (e.g. make it a template for the value type, allow start
to default from a non-zero value) but we do not need that level of
complexity.