Return the element table used by XSPEC
Rather than provide a per-element routine, this just returns a dict
with all elements (indexed by name). We also add a routine to return
the number of elements.

Internally this adds the irange template class to try and use
C++ iterators to replace a simple for(i=0; i<N; ++i) statement.