Add read access to the abundances file used by XSPEC
The low-level API separates out the path from the file, but
the user API combines the two, making use of pathlib.Path. This
routine is not exported, as it's unlikely to be used often.

The direct access naming is a bit confusing here, as for the
path I re-use the "xspath" scheme used for manager/model paths,
but the actual file name doesn't use this.

We could also support changing the path or filename, but it's not
obvious what the logic here is, in that changing the names does not
cause the data to be re-loaded.

Tweak low-level access to the abundance path.

There is currently no-way to access the cross-sections file as is
done for the abundances. However, the access to the abundances can
be made to better-fit the rest of the API (although at present it
looks like we can not change the abundance path from the default,
where it's set to the manager path).