xspec: read access to the abundances file used by XSPEC
We can now send an array of abundances to the C++ layer
rather than having to write the data to a file. This does
not change the UI layer (at least, for users of the
sherpa.astro.xspec module) but it simplifies some of the
internals of that module. In particular, it makes the
state handling a lot-more robust, as we can record the
actual values set, and so re-create them if necessary.

The get_xsabundances_path routine has been added: this is
a simplification of the low-level API, which seperates the
directory and file name, but for users of sherpa.astro.xspec
it does not seem worth maintaining that separation. This
routine is not automatically exported, as it is assumed to
be rarely used. The naming scheme here is a bit confusing:
does path mean a generic path or a Path object.

We could also support changing the path or filename, but it's not
obvious what the logic here is, in that changing the names does not
cause the data to be re-loaded.

There is currently no-way to access the cross-sections data as is
done for the abundances.