abund: can send in an array of abundances
There are multiple ways of doing this, but I decided to overload
set_xsabund so that it can be sent

  - a name (e.g. "wilm")
  - a file name (just as a string)
  - an array of numbers

In doing so I found out that get_xsabund would crash the system if
set_xsabund("file") had been called but no abundances had been
loaded. To avoid this we now make sure that you can not select the
"file" abunbdance table if no data has been loadad.

The "vector" of abundances (whether set directly or from a file)
or stored so that they can be saved/restored with get_xsstate and
set_xsstate.