abund: query by atomic number
As well as providing "direct" access to the element abundances,
the access via element name has been re-worked so that we do not
have to re-direct the err stream during the call. The price for
this is manually checking the name-to-index step, which relies
on some low-level code, but it is less messy than the stream
redirection it replaces.