XSPEC: allow spectrumNumber to be sent to XSPEC models
The XSPEC models now support the keyword argument spectrumNumber
which lets the user send in the spectrumNumber/ifl argument, so
that models like smaug, polconst, pollint, and polpow can access
the necessary XFLT data. This could have used ifl instead of
spectrumNumber, since that's the "older" term from XSPEC, but
I felt it better to use the more-descriptive term, even if
significantly longer, as it is not likely to be something users
are going to type in often (although this partly depends on the
following paragraph).

However, there is currently no way to use this elsewhere in the
system (such as in a fit or when creating a model plot). It also
isn't being passed through certain model combinations (e.g.
XSPEC convolution models) or supported by non-XSPEC models.

Note that the current behavior with the cache code is not ideal,
since a call like

    y1 = mdl(elo, ehi, spectrumNumber=1)
    y2 = mdl(elo, ehi, spectrumNumber=2)

will use the cache (i.e. y2 is the same as y1) even when it should
not be. However, we currently have no XSPEC models that actually
make use of the spectrumNumber argument, so it is not a real
problem. A test has been added to catch when this behavior changes.