cache: better handling of arguments (positional and keyword)
Improve the handling of the cache arguments for a 1D model

- named vs positional (in particular the xhi argument)
- include keyword arguments in the cache

The way we do this means that mdl(x, xhi) and mdl(x, xhi=xhi) will
not be recognized as the same call, but this change in calling
behavior is not going to happen for the cases we want to cache.

This could be a pessimisation for most XSPEC models (well, all
XSPEC models that we currently support) as it means that if the
spectrumNumber kwarg is set then it becomes a key used in the
cache, but at the moment we do not make use of this feature so
we leave the code as is. We may decide to make the cache a bit
smarter (e.g. have an XSPEC-specific version) in the future.