sherpa_test: ensure pytest-xvfb is installed
Ensure that the xvfb pytest plugin is installed before running
the sherpa_test tests. This means that if a user has DS9 installed
they will not see DS9 windows pop up during a test. If the user
wishes to see the windows appear then they can add the

  --no-xvfb

argument to the sherpa_test call, or run the tests directly with

  pytest --pyargs sherpa

Note that there is some issue in the code for when we have to
install pytest-xvfb as part of the test process: should we pass
the new module through to pytest.main? The existing code suggests
we should, as the module will not have been registered correctly.
However, tests shows that the code fails if done like this, so
we do not do this. The import code has not been used for a while
so it's not clear if this is one or more of

- python version change
- pip change
- underlying build infrastructure change
- something odd with pytest-catchlog when it was available
- ?