tablemodel: improve validation
Address some of the issues found in testing and make some small
additions to the documentation.

There are two cache-related bug-fixes:
  - calling load clears the cache
  - setting method clears the cache

In both cases we could have added an optimisation to only clear
the cache if the value has not changed, but it did not seem to
be worth it.

The method attribute is now called-out separately, with its
own docstring, rather than being hidden away. This interacts with
the cache fix by making method a property so we can easily add
documentation and track when it changes.