tablemodel: remove explicit state managenment
The __setstate__ method had logic for restoring "non-standard"
settings, which were presumably to allow old models to be restored.
This code has been in git since the first commit, and so is
unlikely to be useful any more (there are more-extensive changes
to the Sherpa state that have happened since). The code has
therefore been removed. Note that the existing tests - which
check you can pickle and restore a TableModel - still pass.