Tests: add --skipds9 option
The DS9 tests can now be skipped with the --skipds9 option:

    pytest sherpa -n auto --skipds9

The reason for this option is to allow the tests to be run in
parallel, since the DS9 tests do not run successfully in this
case. I have spent some time working out how to get the DS9
tests to run in this case, but even using a xdist_group marker
(added in pytest-xdist 2.5.0) did not seem to work. This is a
significantly-simpler solution. It is only really valid for
evaluating code changes that do not change the DS9 interface,
but that is still a large number of cases.

The underlying problem is that if DS9 is installed on your system
then the DS9 tests will be run for you, and until now there's been
no way to stop this.