XSPEC: add types to the utils code
There are several inter-related changes:

- use dataclass rather than namedtuple.
- tweak class definitons to remove un-needed tokens (e.g.
  `class Foo():` -> `class Foo:`)
- add typing annotations

The typing annotations try to be pragmatic - for instance it
looks like sensibly typing the modelfile argument to
parse_xspec_model_description is too hard to do, so we do not
do so. There are also places where we could be more-specific
but it would complicate the typing code for no significant
benefit in code clarity.

As the documentation refers to these types - either explicitly
or via the typing rules - they have been added to the RTD
documentation.