Note when a model is changed to F77 form (double precision)
The model "name" stored in the model.dat file includes
information on the interface type of the model, in a form
of Hungarian notation [1], but the parse_xspec_model_description
routine strips this out, storing the information as a free-form
string in the language field of the model (I should really
have used an ENUM but decided to use a string). We already
recognized this when checking C++ models, but did not for the
F77 models (FORTRAN double precision). We now do. This is
relevant as XSPEC 12.12.0 (I think) changed three models to
this form and it would be nice to have been told about this,
as we need to write code to support these models.

[1] https://en.wikipedia.org/wiki/Hungarian_notation