Ensure the mask is updated when the grouping is (fix #1160)
When changing the grouping field we need to

  - store the old filter
    BUT only if the data is partially masked
    AND we are not within the __init__ call
  - change grouping
  - clear the filter
  - reapply the old filter

There are several routines that do this, but is there any
comonality to them to exploit?

Note that we only reset the filter if an existing filter has
been applied

a) to save time
b) to retain the logic that says that if mask is True then no
   filter has been applied, so that notice(lo, hi) will be
   treated as "just select the range lo-hi"

If we didn't do this then a mask of True would get changed to
a mask of [True] * ngroup, but then a call like notice(lo, hi)
would be ignored, as the system would think that - because mask
is an iterable - a filter had been applied when it had not.