Tests: corner case of grouping code
Ensure we test a fall-thruogh case in the grouping code and
add a test of issue #1636 found whilst working on the code.

There are also changes to tests to prepare for the changes
to the ui group commands, and some changes added to improve
test coverage (related to changes made in this PR set).