utils: rework access to PHA data (internal)
Rework code that looks something like

    d = self._get_pha_data(id)
    if bkg_id is not None:
        d = self.get_bkg(id, bkg_id)

(or the more-recent version where there's a if .. else .. form)
by moving the logic for this into _get_pha_data.

There is still repeated code where we don't require a PHA dataset
when bkg_id is None and this could do with a similar change. I
have tried to move these examples to the same form

    if bkg_id is None:
        d = self.get_data(id)
    else:
        d = self.get_bkg(id, bkg_id)

so it is easier to find if we ever come up with a replacement.

There is some clean-up code to use idval rather than id as a
variable name - but as id is part of the API it is implemented as

    idval = self._fix_id(id)

This is only really done for those routines that may need to output
the idval (e.g. in an error message), as  get_data/get_bkg call
_fix_id themselves if needed (and it's not obviously helpful to
cache this check, as it doesn't  really save much time). Some
routines require a _fix_id call because we then access dictionaries
with this value,  rather than going through a method.

There is also one clean up removing an unused variable - setting
err=None - as we may as well clean this up while we can.