utils: rework access to PHA data (internal)
Rework code that looks something like

    d = self._get_pha_data(id)
    if bkg_id is not None:
        d = self.get_bkg(id, bkg_id)

(or the more-recent version where there's a if .. else .. form)
by moving the logic for this into _get_pha_data.

There is still repeated code where we don't require a PHA dataset
when bkg_id is None and this could do with a similar change. I
have tried to move these examples to the same form

    if bkg_id is None:
        d = self.get_data(id)
    else:
        d = self.get_bkg(id, bkg_id)

so it is easier to find if we ever come up with a replacement.