utils: rework access to PHA data (internal)
Rework code that looks something like

    d = self._get_pha_data(id)
    if bkg_id is not None:
        d = self.get_bkg(id, bkg_id)

(or the more-recent version where there's a if .. else .. form)
by ,oving the logic for this into _get_pha_data.

There is still repeated code where we don't require a PHA dataset
when bkg_id is None and this could do with a similar change.