utils: remove potentially-dangerous API use
Setting a parameter to a mutable field - e.g. extra_ids={} - is
potentially dangerous and pylint flags it as such. Fortunately
we can convert it in these cases to setting the default for the
parameter to None, since we don't rely on the "dangerous behavior"
here (and these are internal routines which are not expected to be
used outside of Sherpa).