Tests: ensure corner cases are covered
Note that this includes test cases that currently fail (but will
be fixed soon): #1641 and #1642

A number of tests are labelled with the skip_if_no_io decorator
which skips a test for the AstroSession case when there's no I/O
module available. Prior to this we skipped the test even when
using the Session class, as I hadn't been able to come up with
a "nice" way to achieve this. There probably is a better way to
do this with pytest.param and the marks argument, but the main
issue is that sherpa.utils.testing doesn't make it easy to get
at the test that the decorators like requires_fits runs, so we
end up having to replicate this code here. We could re-work the
testing module, but I do not think it's worth it here.