bugfix: fix #1641, fix #1642
Issue #1641

Since ignore() calls notice() then we don't need the wrapper
around the ignore superclass and can just add the extra logic
into notice. This logic doesn't therefore need to be in a separate
routine but can be done in situ, and we can use a set as that is
semantically better than the old check.

The warning message is augmented with the list of the different
analysis settings, as that may prove useful to users in this
presumably-rare situation.

Issue #1642

Fix a simple typo introduxed during a previous round of lint-style
checks (#175)