utils: wrapped models now reference the wrapped model fix #215
Previousy help(gauss1d) in Sherpa would not provide anything
helpful. Now it at least mentions the Gauss1D model that is
wrapped, including the first line of the model's docstring,
if set.

It is not clear what will really help the user here,
but this has to be better than the WrappedModel text. See
issue #215 for more information.