Tests: add unexpected behavior with group_snr (see #1646)
Working on PR #1637 I noticed that group_snr could change the
filter, so I have added a test showing the issue and also some
low-level tests showing the underlying problem noted in the
issue #1646. It looks like the group library has a bug for the
snr case. As we treat the group code as an external requirement,
and may not even build it (for CIAO) there's not a lot we can do
to address this issue. The tests should show if the underlying
group code gets fixed.